radar fixes
