+++ tests/dummy/app/screens/examples/infinite-scroll/template.hbs
@@ -3,7 +3,7 @@
         <div class="col-sm-5">
             <h3>Demo</h3>
             <div class="table-wrapper dark">
-              {{!- BEGIN-SNIPPET infinite-scroll-example }}
+              {{!- BEGIN-SNIPPET smart-collection-example }}
               {{#smart-collection
                 model.numbers
                 itemDimensions=defaultDimensions
@@ -36,7 +36,7 @@
               they return.
           </p>
             <h4>Code for Demo</h4>
-          {{code-snippet name="infinite-scroll-example.hbs"}}
+          {{code-snippet name="smart-collection-example.hbs"}}
         </div>
     </div>
 </div>
